fix(auth): Copy-only, best-effort legacy device metadata recovery
Address review feedback:
- Keep the legacy entry instead of deleting it: sign-in paths still read
  device metadata by the typed alias, so deleting it would make the next
  sign-in register a duplicate device
- Make the copy best-effort: return the metadata even if persisting the
  copy fails, instead of failing a refresh we can serve
- Restore the 'Invalid Refresh Token.' symptom in the KDoc
- Fold the alias claim tests into TokensTest and drop the Robolectric
  runner they did not need
