+++ aws-auth-cognito/src/test/java/com/amplifyframework/auth/cognito/actions/FetchAuthSessionCognitoActionsTest.kt
@@ -295,7 +295,6 @@ class FetchAuthSessionCognitoActionsTest {
             }
         }
         coJustRun { credentialStoreClient.storeCredentials(any(), any()) }
-        coJustRun { credentialStoreClient.clearCredentials(any()) }
 
         val requestSlot =
             slot<aws.sdk.kotlin.services.cognitoidentityprovider.model.GetTokensFromRefreshTokenRequest>()
